refactor: move ConditionWithType to top level
