+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/dependent/workflow/WorkflowResult.java
@@ -104,7 +104,7 @@ class WorkflowResult {
     }
 
     DetailBuilder<R> withResultForCondition(
-        DependentResourceNode.ConditionWithType conditionWithType,
+        ConditionWithType conditionWithType,
         ResultCondition.Result conditionResult) {
       switch (conditionWithType.type()) {
         case ACTIVATION -> activationConditionResult = conditionResult;
